[#162] refactor: repository 패키지를 적절한 곳으로 이동
