+++ src/main/java/goorm/eagle7/stelligence/domain/contribute/scheduler/DebateHandler.java
@@ -5,7 +5,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import goorm.eagle7.stelligence.domain.contribute.ContributeRepository;
 import goorm.eagle7.stelligence.domain.contribute.model.Contribute;
-import goorm.eagle7.stelligence.domain.debate.DebateRepository;
+import goorm.eagle7.stelligence.domain.debate.repository.DebateRepository;
 import goorm.eagle7.stelligence.domain.debate.model.Debate;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
