+++ src/main/java/goorm/eagle7/stelligence/domain/debate/repository/CommentRepository.java
@@ -1,4 +1,4 @@
-package goorm.eagle7.stelligence.domain.debate;
+package goorm.eagle7.stelligence.domain.debate.repository;
 
 import java.util.List;
